Rename Pointers in C to Pointers in  C
Complete the function void update(int *a, int *b). It receives two integer pointers, int* a and int* b. Set the value of a to their sum, and b to their absolute difference. There is no return value, and no return statement is needed.
a'=a+b
b'=|a-b|

Sample Input
4
5
Sample Output
9
1